+++ jabgui/src/main/java/module-info.java
@@ -98,10 +98,7 @@ open module org.jabref {
     // region: SQL databases
     // requires embedded.postgres;
     // requires org.tukaani.xz;
-    // requires ojdbc10;
     // requires org.postgresql.jdbc;
-    // requires org.mariadb.jdbc;
-    // uses org.mariadb.jdbc.credential.CredentialPlugin;
     // endregion
 
     // region: Apache Commons and other (similar) helper libraries